requirements.txt
discord.py
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
awscli==1.22.8
bcdoc==0.12.2
botocore==1.23.8
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.2.5
discord.py==1.7.3
docutils==0.15.2
idna==3.3
jmespath==0.10.0
multidict==6.0.2
Pillow==9.1.0
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.27.1
rsa==3.1.2
s3transfer==0.5.0
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.7
yarl==1.7.2